pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solar-panel-fault-detection"
version = "1.0.0"
description = "Production-grade solar panel fault detection platform with SparkNet CNN and U-Net"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Solar Detection Team", email = "team@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["solar", "fault-detection", "deep-learning", "cnn", "computer-vision"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "albumentations>=1.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.28.0",
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
    "httpx>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.66.0",
    "xgboost>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[project.scripts]
train-sparknet = "scripts.train_sparknet:main"
train-unet = "scripts.train_unet:main"
evaluate = "scripts.evaluate:main"
serve = "scripts.serve:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "config*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "alembic"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["alembic", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
